requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
python-magic==0.4.27
yara-python==4.3.1
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-wtf==1.1.1
flask-bootstrap==3.3.7.1
clamd==1.0.2
GitPython>=3.1.24
starlette==0.27.0